requirements.txt
Flask==1.1.2
pylint==2.7.2
twilio==6.53.0
python_dotenv==0.15.0
Flask_WTF==0.14.3
